+++ libs/docs/core/select/examples/select-forms/select-forms.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
+import { ButtonComponent } from '@fundamental-ngx/core/button';
 import { SelectModule } from '@fundamental-ngx/core/select';
 
 @Component({
@@ -7,7 +8,7 @@ import { SelectModule } from '@fundamental-ngx/core/select';
     templateUrl: './select-forms.component.html',
     styleUrls: ['select-forms.component.scss'],
     standalone: true,
-    imports: [FormsModule, ReactiveFormsModule, SelectModule]
+    imports: [FormsModule, ReactiveFormsModule, SelectModule, ButtonComponent]
 })
 export class SelectFormsComponent {
     options: string[] = ['Apple', 'Pineapple', 'Tomato', 'Strawberry'];
@@ -19,4 +20,10 @@ export class SelectFormsComponent {
     disabledForm = new FormGroup({
         disabledControl: new FormControl({ value: 'Apple', disabled: true }, Validators.required)
     });
+
+    toggleDisabledState(): void {
+        this.disabledForm.controls.disabledControl.enabled
+            ? this.disabledForm.controls.disabledControl.disable()
+            : this.disabledForm.controls.disabledControl.enable();
+    }
 }