+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.47.4-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.5...v0.47.4-rc.6) (2023-11-23)
+
+
+### Bug Fixes
+
+* **core:** segmented button a11y improvements ([#10933](https://github.com/SAP/fundamental-ngx/issues/10933)) ([72c6c71](https://github.com/SAP/fundamental-ngx/commit/72c6c71f6edf960a8b430595f46ab86677e7620c))
+
+
+
+
+
+## [0.47.4-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.4...v0.47.4-rc.5) (2023-11-22)
+
+
+### Bug Fixes
+
+* **core:** select disabled state ([#10994](https://github.com/SAP/fundamental-ngx/issues/10994)) ([e41be1a](https://github.com/SAP/fundamental-ngx/commit/e41be1a672307019fcbf4c7abdbcde255ece699b))
+
+
+
+
+
 ## [0.47.4-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.3...v0.47.4-rc.4) (2023-11-20)
 
 **Note:** Version bump only for package fundamental-ngx